mtapi_network_c: removed visual studio warning

+++ mtapi_plugins_c/mtapi_network_c/test/embb_mtapi_network_test_socket.cc
@@ -53,7 +53,7 @@ void NetworkSocketTest::TestBasic() {
 
   err = embb_mtapi_network_socket_initialize(&server_sock);
   PT_EXPECT(err != 0);
-  int port = 4700;
+  uint16_t port = 4700;
   do {
     port++;
     err = embb_mtapi_network_socket_bind_and_listen(
